requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
importlib-metadata==3.7.3
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.3
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver @ file:///D:/bld/kiwisolver_1610099949969/work
MarkupSafe==1.1.1
matplotlib @ file:///D:/bld/matplotlib-suite_1605180495775/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.3
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
packaging==20.9
pandas @ file:///C:/ci/pandas_1602088205210/work
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow @ file:///D:/bld/pillow_1610407628862/work
pipreqs==0.4.10
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.15.6
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.7.4
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.0
requests==2.25.1
scikit-learn @ file:///C:/ci/scikit-learn_1598376983131/work
scipy @ file:///C:/ci/scipy_1597686737426/work
seaborn @ file:///tmp/build/80754af9/seaborn_1600553570093/work
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
streamlit==0.79.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.11.1
tornado @ file:///D:/bld/tornado_1610094878337/work
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
yarg==0.1.9
zipp==3.4.1
